requirements.txt
streamlit
sentence-transformers
PyPDF2
python-docx
pandas